+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## Release 2.2.0a2 (WIP)
+## Release 2.2.0a4 (WIP)
 
 ### Breaking Changes
 - Removed `gym` dependency, the package is still required for some pretrained agents.
 
 ### New Features
+- Add `--eval-env-kwargs` to `train.py` (@Quentin18)
 
 ### Bug fixes
 
@@ -13,6 +14,7 @@
 - Updated docker image, removed support for X server
 - Replaced deprecated `optuna.suggest_uniform(...)` by `optuna.suggest_float(..., low=..., high=...)`
 - Switched to ruff for sorting imports
+- Updated tests to use `shlex.split()`
 
 ## Release 2.1.0 (2023-08-17)